move modes tuning to dmd_modes_tuner.py, improve docs, new tests
