+++ docs/source/dmd_modes_tuner.rst
@@ -1,0 +1,16 @@
+DMD Modes Tuner
+=================
+
+.. currentmodule:: pydmd.dmd_modes_tuner
+
+.. automodule:: pydmd.dmd_modes_tuner
+
+.. autosummary::
+	:toctree: _summaries
+	:nosignatures:
+
+	select_modes
+	stabilize_modes
+
+.. autoclass:: ModesSelectors
+	:members:
