+++ tests/test_dmdbase.py
@@ -105,129 +105,6 @@ class TestDmdBase(TestCase):
         dmd = DMDBase(sorted_eigs='real')
         assert dmd.operator._sorted_eigs == 'real'
 
-    def test_select_modes(self):
-        def stable_modes(dmd_object):
-            toll = 1e-3
-            return np.abs(np.abs(dmd_object.eigs) - 1) < toll
-        dmd = DMD(svd_rank=10)
-        dmd.fit(sample_data)
-        exp = dmd.reconstructed_data
-        dmd.select_modes(stable_modes)
-        np.testing.assert_array_almost_equal(exp, dmd.reconstructed_data)
-
-    def test_stable_modes_both(self):
-        class FakeDMD:
-            pass
-
-        fake_dmd = FakeDMD()
-        setattr(fake_dmd, 'eigs', np.array([1 + 1e-4, 2, 1 - 1e-2, 5, 1, 1 - 0.5*1e-3]))
-
-        expected_result = np.array([False for _ in range(6)])
-        expected_result[[0, 4, 5]] = True
-
-        assert all(DMDBase.ModesSelectors.stable_modes(max_distance_from_unity=1e-3)(fake_dmd) == expected_result)
-
-    def test_stable_modes_outside_only(self):
-        class FakeDMD:
-            pass
-
-        fake_dmd = FakeDMD()
-        setattr(fake_dmd, 'eigs', np.array([1 + 1e-4, 2, 1 - 1e-2, 5, 1, 1 - 0.5*1e-3]))
-
-        expected_result = np.array([False for _ in range(6)])
-        expected_result[[0, 2, 4, 5]] = True
-
-        assert all(DMDBase.ModesSelectors.stable_modes(max_distance_from_unity_outside=1e-3)(fake_dmd) == expected_result)
-
-    def test_stable_modes_inside_only(self):
-        class FakeDMD:
-            pass
-
-        fake_dmd = FakeDMD()
-        setattr(fake_dmd, 'eigs', np.array([1 + 1e-4, 2, 1 - 1e-2, 5, 1, 1 - 0.5*1e-3]))
-
-        expected_result = np.array([False for _ in range(6)])
-        expected_result[[0, 1, 3, 4, 5]] = True
-
-        assert all(DMDBase.ModesSelectors.stable_modes(max_distance_from_unity_inside=1e-3)(fake_dmd) == expected_result)
-
-    def test_stable_modes_errors(self):
-        with self.assertRaises(ValueError):
-            DMDBase.ModesSelectors.stable_modes()
-        with self.assertRaises(ValueError):
-            DMDBase.ModesSelectors.stable_modes(max_distance_from_unity=1.e-2, max_distance_from_unity_inside=1.e-3)
-        with self.assertRaises(ValueError):
-            DMDBase.ModesSelectors.stable_modes(max_distance_from_unity=1.e-2, max_distance_from_unity_outside=1.e-3)
-
-    def test_threshold(self):
-        class FakeDMD:
-            pass
-
-        fake_dmd = FakeDMD()
-        setattr(fake_dmd, 'eigs', np.array([complex(1, 1e-4), 2, complex(1, 1e-2), 5, 1, complex(1, 5*1e-2)]))
-
-        expected_result = np.array([False for _ in range(6)])
-        expected_result[[1, 5]] = True
-
-        assert all(DMDBase.ModesSelectors.threshold(1+1.e-3, 2+1.e-10)(fake_dmd) == expected_result)
-
-    def test_compute_integral_contribution(self):
-        np.testing.assert_almost_equal(DMDBase.ModesSelectors._compute_integral_contribution(
-            np.array([5,0,0,1]), np.array([1,-2,3,-5,6])
-        ), 442, decimal=1)
-
-    def test_integral_contribution(self):
-        class FakeDMD:
-            pass
-
-        fake_dmd = FakeDMD()
-        setattr(fake_dmd, 'dynamics', np.array([[i for _ in range(10)] for i in range(4)]))
-        setattr(fake_dmd, 'modes', np.ones((20, 4)))
-        setattr(fake_dmd, 'dmd_time', None)
-        setattr(fake_dmd, 'original_time', None)
-
-        expected_result = np.array([False for _ in range(4)])
-        expected_result[[2, 3]] = True
-
-        assert all(DMDBase.ModesSelectors.integral_contribution(2)(fake_dmd) == expected_result)
-
-    def test_integral_contribution_reconstruction(self):
-        dmd = DMD(svd_rank=10)
-        dmd.fit(sample_data)
-        exp = dmd.reconstructed_data
-        dmd.select_modes(DMDBase.ModesSelectors.integral_contribution(2))
-        np.testing.assert_array_almost_equal(exp, dmd.reconstructed_data)
-
-    def test_stabilize_modes(self):
-        class FakeDMD:
-            pass
-
-        class FakeDMDOperator:
-            pass
-
-        dmd = DMD()
-        fake_dmd_operator = FakeDMDOperator()
-
-        eigs = np.array([complex(0.3, 0.2), complex(0.8,0.5), 1, complex(1,1.e-2), 2, complex(2,1.e-2)])
-        amplitudes = np.array([1,2,3,4,5,6], dtype=complex)
-
-        setattr(fake_dmd_operator, '_eigenvalues', eigs)
-        setattr(fake_dmd_operator, 'eigenvalues', eigs)
-        setattr(dmd, '_Atilde', fake_dmd_operator)
-
-        setattr(dmd, '_b', amplitudes)
-
-        dmd.stabilize_modes(0.8, 1.2)
-
-        np.testing.assert_array_almost_equal(
-            dmd.operator._eigenvalues,
-            np.array([complex(0.3, 0.2), complex(0.8,0.5) / abs(complex(0.8,0.5)),
-                1, complex(1,1.e-2) / abs(complex(1,1.e-2)), 2, complex(2,1.e-2)]))
-
-        np.testing.assert_array_almost_equal(
-            dmd._b,
-            np.array([1, 2*abs(complex(0.8,0.5)), 3, 4*abs(complex(1,1.e-2)), 5, 6]))
-
     def test_enforce_ratio_y(self):
         dmd = DMDBase()
         supx, infx, supy, infy = dmd._enforce_ratio(10, 20, 10, 0, 0)
